requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.17.0
requests==2.31.0
py5paisa==0.7.2
black==23.7.0
pyotp==2.9.0
websockets==11.0.3
pymongo==4.4.1
python-telegram-bot==20.4
python-telegram-bot[job-queue]==20.4
redis==5.0.0
mkdocs==1.5.2
reorder-python-imports==3.10.0
